Basic implementation of the Services Registry concept for Morpho.  Now the client framework allows plugins to register the services that they support by name, and then those services can be invoked by passing a ServiceRequest object (containing the data) to ClientFramework.  The ClientFramework will then search its service registry and, if an appropriate service has been registered, will pass the request to that service to be handled.  This is a form of loose binding (ala RMI/CORBA) that is lightweight (because it doesn;t require a special dispatcher) but also doesn;t have all of the features of those services (e.g., object inspection is missing).  This capability allows us to create a series of plugins that can interact loosely and pass data between them without knowing a compile time which services will be present.
Of course, each plugin will have dependencies on services from other plugins
that must be present in order for the plugin to function correctly.  We need
to add a method to ClientFramework that allows the plugin to inspect the
available services to be sure all ofits requirements are met at runtime.
This basically means that plugins are establishing mutual services that
are identified only by name; the serviceName represents a behavioral contract
that must be met if we expect the plugins to continue to function.

Also finished cleaning up the menu and toolbar additions, etc.