Added new build targets for running morpo, and running JUnit tests.
Right now the JUnit tests are only working if you run them through the "junit"
target (which manually launches the TestRunner application.  Ant is supposed to
be able to handle Junit tests with the "junit" task (as exemplified in my
"test" target, but this currently does not work for ClientFramework because of an eror calling Class.getResource().  I suspect this error is related to an
incorrectly assigned, or at least different, class loader that can't find the
resources in question (image files).