Developed a unit test (see literature on extreme programming) that works under the JUnit testing framework to test the new ClientFramework functions for logging into and out of Metacat.
This requires that you download JUnit from http://www.junit.org.  I used
junit-3.6.

Running "ant test" or "ant junit" should
run this test and produce no failures if the connection code to
metacat is working properly.  Note that you'll need to provide a correct
username and password for the tests to succeed (we need a better way to pass
data to these tests, especially confidential data, without hardcoding it in the
test -- there are several possible open-source solutions for this that I
will explore).