Added new classes to the query package to allow the separation of user interface from data manipulation.  The ResultSet class now stores the results of queries, and is working for Metacat queries. The Query class represents a query, and is composed of QueryGroups and QueryTerms. The current interface shows an "owner" query, by default for the user berkley (because he has docs in Metacat and so I put his username in config.xml). If you log in and no docs show up in the main window, its probably because you don't have any metadata that you own in Metacat.
Still need to customize the resultset to display a more reasonable set
of columns, and the butterfly icon is just a placeholder for future
network/local icons.

Need to add back in the local query functionality, and make the
Query.execute() method merge results from the metacat query and local query.
Also need to reestablish the coed that determines whether a local or
network query is done (probably should be attributes of the Query class, and
persistent in config.xml).

Still need to nix the tabbed pane GUI layout in ClientFramework.

Lots more still to do :-) but this was fun.