Cleaned up QuertPlugin, removed dead code, added documentation.
Made the ownerQuery dunamically configure itself based on the username in
the config file and the proper returndocs in the config file. Right now the
getOwnerQuery() method constructs the query by making an XML string, but it
should be changed to use the Query setter methods and let Query
generate the text version for Metacat.  However, because that part of Query is
not yet implemented, I am just passing a text string for now.

Cleaned up ResultSet, added documentation, removed dead code.