Changed the framework window so that it no longer uses a tabbed pane. Wrote additional documentation, cleaned up code, removed dead code.
Revised SubmitDataDialog and SubmitDialog so that they don't refer to
protected variables in the framework (in particular, openFileDialog).  But now
these two classes are non-functional -- as there not in use currently, I
thought that was ok.  In general, we should not be referencing member
variables across classes.  Instead, write an accessor method so that our
implementations can change.  All member variables should be private!