Added new class ResultFrame that is used to display results of queries that are generated by the QueryDialog.  This frame contains an embedded ResultPanel that actually does the display.  The frame automatically registers itself with the framework on creation, and removes itself upon closing so that the framework's Window menu stays up to date.  The Query title is used for the window title.
Created some test code in QueryPlugin that exercises this feature by
running a query for the "owner" when the "Search..." action is
fired.  This will be replaced by the QueryDialog when it is
finished.

Still need to add "Refresh" and "Revise" functions to the ResultPanel
class so that the queries can be updated.