Added Refresh function to ResultPanel, and new buttons for Refresh and Revise.
In the process, I created a new ConnectionListener interface that allows
classes to be notified of changes to the connection status and changes in the
username.  This is important so that panels like "My Data" will refresh with
data from the right user when the username changes, and get user-private data
when the connection changes to a connected status.  There is a new
method ClientFramework.addCOnnectionListener() that is used to register
listeners withthe framework for these notification events.  See the example in
QueryPlugin to see how it works and is used.