Added new jar file so that we have access to all of the Java Look and Feel icons.  Changed all references to icons to point to the new versions in the jar file (so, if you need an icon for save, save as, new, open, or anything else remotely standard, look in the lib/jlfgr-1_0.jar before adding any new icons.
Refined the UI of ResultSet handling by providing the merge capability
between two set of results (one local, one from metacat). There is a
different icon for result records that exist only on metacat, only locally,
and in both places. Had to modify LocalQuery to make the row Vector
match the row vector produced in ResultSet.  Still have a mismatch
between the data structures used to store triples in LocalQuery and
ResultSet -- this needs to be resolved to use only one structure.

Modified the way in which QueryDialog is called so that it can be used
to revise a query (in addition to its current ability to generate a new
query).  Added accessor methods that allow access to the QueryDialog's
query.  More work to be done here.