Further work on the new profiles feature of morpho. Now, upon morpho startup, we check to see if there was a profile in effect from a previous session, and if so, we load that profile and ask the user to log in.  If not, we prompt the user to create a new profile, which initializes the profiles directory and creates a default profile.  The "profiles/username" directory contains the profile file (username.xml) and the "data", "cache", and "temp" directories for that user. Any user-specific data should now be stored in their profile directory.
Please note that only global configuration information is now
stored in "lib/config.xml" (although I haven't excised it all yet),
and that user specific configuration
information is now stored in "profiles/username/username.xml".
The easy (and correct) way to programatically access this
information is to get a handle to the current profile object from
the framework and use it, like this:

  ConfigXML profile = framework.getProfile();
  String datadir = profile.get("datadir", 0);

I have modified DataStore to use these new profile directories,
but have not yet modified LocalQuery to reflect this change.
Will do so tomorrow.  Also found several places in the code
where the "lib/config.xml" is hardcoded and should not be,
so I will fix this as well (by getting the config object from the
framework properly).

TODO
1) Still need to copy sample data to the profile dir when creating the profile.
2) Possibly create a metacat account if it doesn't exist
3) Change behaviour to deal properly with the LDAP baseDN
4) Change behaviour to properly deal with unique scope for accession #
5) Decide what info we need for the profile and finish collecting and
   storing it in the profile (e.g., contact info)