Modifications to the profile handling code to make the profile & login sequence more efficient.
Modified LocalQuery to now use the data found in a user's profile directory
instead of the 'XMLWorkfolder'.  This means that the static buildPackageList()
method can't work any more (because profile information isn't available unless
the framework (and thus profile info) have been instantiated).  So I removed
it.  Also, changed the key on the various caches in LocalQuery to all use
the docid as the key, rather than various filenames, because it is simpler
and in line with our supposition that all accession numbers must be unique.

You'll need to delete your profile directory and reload it to see the
profile changes that I have made. And also get a clean copy of lib/config.xml.