1) Added new feature that allows plugins to add a separator following a menu when they call "addMenu()" -- previously they could only have a separator before.
2) Fixed the getNextId() function so that it properly returns the next
id in the series.  This was broken because I moved the id numbers into
profiles, but the profile didn't save the new id.  Changed the method
to synchronized just for safety.

3) Extended the ClientFramework.debug(int, String) method to use
dialog boxes for severe errors. Messages with severity 1 to 4 will
result in an error dialog box for the user to inspect.  Those with
severity 5-9 result in a warning dialog box for the user to inspect.
Those with severity greater than 9 are printed only to standard error.
This change means that we should be careful with our wording of errors
that are at a high severity level because the user will see them.
Also be sure to format them reasonably with \n at about 30 chars. A
possible extension of this is to have the application exit on true errors
(severity 1-4) after notifying the user, with the rationale that these
represent design flaws and are so unexpected that there is a chance for
data loss/corruption.