Fixed bug in ConfigXML where options were not being set correctly if there is already an empty element in the file.
Added new feature to ProfileDialog that lets the user use an existing profile
if they type in a name that already exists.  This way, if the reference
to "current_profile" is lost from config.xml, it can easily be re-established
without recreating the profile directory.

Delayed the close of the splash frame in ClientFramework to accomodate
longer startup times.