Completed work on the "Saved queries" feature of the query package. See Bug #232 for details.  Now people can save queries, they show up in the search menu, selecting one re-launches the query, which can then be revised and re-saved.  Saved queries are persistent across morpho sessions, and the Search menu is updated when morpho switches from one profile to another.
Also updating config.xml to remove berkley's hardcoded username from
current_profile (again).  Please do not check in copies of config.xml
with a profile name in <current_profile> because this prevents morpho
from working properly for people that checkout a clean copy of morpho
but haven't yet created a profile.  I'm trying to get Mark and Chris
to test it and give feedback, and we need for it to work "out of the
box" for them.  Thanks.