added support for the package editor to update a file that is located both on metacat and locally.
the localQuery GUI needs to be altered so that it only shows the most recent version of a file that has been hit.  right now if there are two ore more revisions, it will display them all in the result list.  Also, when a document hit on metacat and a document hit locally have the same id, they should be displayed in 1 row, not 2.