cat /tmp/cvsoBE3H3
To be consistent with the EML files, I removed the:

<?xml version="1.0" encoding="UTF-8"?>

line that Spy puts into the dtds.  Also, I changed the content model for
the dsgpolyo and dsgpolyx elements in eml-coverage, eml-dataset, and eml-entity
to reflect the same EML changes.

Also, I fixed the typo in the altmax element (was almax).