Updated the morpho module to properly handle JUnit tests.  Now the tests are stored in their own subtree (tests) and are compiled separately from the main Morpho code.  This prevents compilation errors when the junit.jar file can't be found.
Removed the junit.jar file from the lib directory, and instead installed
it as part of ant as we are doing for other modules (it was an old
version anyways). Note that this means you'll need to download and
install junit 3.7 on any machine that you want to build morpho on (just
link it into the ant/lib directory and it'll be found).
This is better than replicating junit.jar through all of our modules
for every version of junit that we use.

Now one can run "ant test" and the tests will be compiled and run.
Please be sure to run "ant test" before you commit code to CVS from now
on.  Also, as you write new code, please wirte a class that tests its
public methods thoroughly.  You can look at the dsparse module for some
examples of how tests can be structured and written.