This is a partial fix for http://bugzilla.ecoinformatics.org/show_bug.cgi?id=1561
I've added in code to handle each of the XML attributes that may be present in
an <attribute> such as id, scope, system, and further down in the tree,
<storageType typSystem="">.

These changes add in String placeholders for each XML attribute.  The
getPageData() and setPageData() methods were changed to get and set these
values from the OrderedMap that is passed in.

I've also fixed a related bug that is caused by calling setPanelData()
for each of the appropriate measurement scales (nominal, ordinal, interval,
ratio, datetime).  Previously, setPanelData would be called for every
measurementScale.  Now, the setPanelData() method is only called for the
measurementScale that was determined by findMeasurementScale().  This allows
the AttributePage dialog to be used rather than the tree editor, which it was
defaulting to above because setPanelData() was called multiple times
consecutively.

In future versions of this dialog, widgets should be used to show the attribute
values if desired.  For now they are just placeholders in the class.