Like the fixes in AttributePage.java, this is a partial fix to http://bugzilla.ecoinformatics.org/show_bug.cgi?id=1561.
I've added a String placeholder for the <enumeratedDomain enforced="">
attribute.  This changes the setPanelData() and getPanelData() methods.

Also, I added in minor documentation.

This fix brings up another bug: <enumeratedDomain> is repeatable, but the
AttributePage class is only set up to handle the first instance of this element
by using the NominalOrdinalPanel, IntervalRatioPanel, etc.  I'll add this in as
a bug.