I've added placeholders for the <missingValueCode> sub elements: <code> and <codeExplanation>.  This changed the setPageData() and getPageData methods.
Also, I corrected a mistake I made with setting the storageTypeField.

This is a partial fix to http://bugzilla.ecoinformatics.org/show_bug.cgi?id=1561

Still need to deal with coverage and method.  This fix also only supports a
single missing value code/codeExplanation pair for the attribute in question.
When this is converted to a widget in the interface, we should handle the case
were missingValueCode elements repeat.