Fixed a bug in geographic screen. Bug reported by Kristin Vanderbilt and bug report at http://bugzilla.ecoinformatics.org/show_bug.cgi?id=2103
The code earlier changed the value of coordinated entered if value of east coordinate was same as value of west coordinate - this was done so that a box can be displayed when Box Tool was selected.
But the code was executed irrespective of whether Box or Point tool was selected.
Fixed it so value is changed only when box tool is selected.