This commit is for fixing bug 1622
http://bugzilla.ecoinformatics.org/show_bug.cgi?id=1622

The changes made in this bug result in the following:
1. The user is notified if there is a field which cant be trimmed because it is required and text has not been provided for that field.
2. The Trim button is faster as some of the redundant computation is removed now.

The above changes should fix some of the problems that users face while using the tree editor - for example, using the trim and show all button results in invalid eml documents.

On the other hand, tree editor code is complex and the above changes might not have fully fixed the problems. More testing is required.