Made changes in displayed text as requested by Veronique below:
1) Morphos main window: text for duplicate icon ("Duplicate this datapackage and save locally..."): replace datapackage by data package.
3) New Profile window/Metacat Account Information:
(a) on right panel, change Metacat Account Information to Network Account Information and delete Network in front of Username;
(b) on left panel change current text to this: Enter the information you submitted when you registered for the Knowledge Network for Biocomplexity (KNB). If you have not registered for the KNB yet, go to http://ldap.ecoinformatics.org/cgi-bin/ldapweb.cgi?cfg=knb. This will allow you to login to the network and collaborate with other researchers through the KNB.