Please note these comments are for this version and parts of the previous one.
The Main Welcome Screen Page has had the screen shots redone to accommodate the changes in Morpho 1.6.0 menu choices. The snapshots show the login button in two Windows views.  (In the Display Properties in Windows>
Appearance tab>Windows and buttons: there are two choices Windows Classic style and Windows XP style). With Windows Classic style the raised buttons for the toolbar and login show and with Windows XP style the login and toolbars appear flush to the surface.

The orignal Welcome Screen Page had a larger window size due to the table settings. Now the window sizes to the largest graphic file, like the other pages.